fix: c_int integers in `print_update_clm`

+++ bldsva/intf_DA/pdaf1_1/model/clm3_5/enkf_clm_mod.F90
@@ -96,7 +96,7 @@ subroutine print_update_clm(ts,ttot) bind(C,name="print_update_clm")
 
     implicit none
 
-    integer, intent(in) :: ts,ttot
+    integer(c_int), intent(in) :: ts,ttot
 
     ! *** local variables ***
     integer :: numg           ! total number of gridcells across all processors
